Added binary output and increased hex input/output to 64 bit
git-svn-id: file:///home/anders/svn/norman/DevCalc.NET/trunk@87 1a7889ac-cf86-4f42-b9eb-dc89fd96a95d

+++ DevCalc.NET/MathParser.cs
@@ -170,7 +170,7 @@ namespace DevCalcNET
             }
             if (val.StartsWith("0x"))
             {
-                return new Constant(int.Parse(val.Remove(0, 2), System.Globalization.NumberStyles.HexNumber));
+                return new Constant(Int64.Parse(val.Remove(0, 2), System.Globalization.NumberStyles.HexNumber));
             }
             else if (val.StartsWith("0b"))
             {
